Extract CompletionFeature from EditorEventManager
